reorganizing launch script
